+++ pygsp/graphs/graph.py
@@ -166,7 +166,7 @@ class Graph(object):
         r"""
         Create a deepcopy of a graph with all the attributes.
 
-        Exemples
+        Examples
         --------
         >>> from pygsp import graphs
         >>> G = graphs.Logo()