+++ pygsp/plotting.py
@@ -480,7 +480,7 @@ def plot_filter(filters, G=None, npoints=1000, line_width=4, x_width=3,
     #     plt.show()
 
 
-def plot_signal(G, signal, show_edges=None, cp={-6, -3, 160},
+def plot_signal(G, signal, show_edges=None, cp=[-6, -3, 160],
                 vertex_size=None, vertex_highlight=False, climits=None,
                 colorbar=True, bar=False, bar_width=1, savefig=False,
                 plot_name=None):
@@ -542,7 +542,7 @@ def plot_signal(G, signal, show_edges=None, cp={-6, -3, 160},
     if climits is None:
         cmin = 1.01 * np.min(signal)
         cmax = 1.01 * np.max(signal)
-        climits = {cmin, cmax}
+        climits = [cmin, cmax]
     if plot_name is None:
         plot_name = "Signal plot of " + G.gtype
 
@@ -615,7 +615,7 @@ def plot_signal(G, signal, show_edges=None, cp={-6, -3, 160},
     #     plt.show()
 
 
-def pg_plot_signal(G, signal, show_edges=None, cp={-6, -3, 160},
+def pg_plot_signal(G, signal, show_edges=None, cp=[-6, -3, 160],
                    vertex_size=None, vertex_highlight=False, climits=None,
                    colorbar=True, bar=False, bar_width=1):
     r"""
@@ -660,7 +660,7 @@ def pg_plot_signal(G, signal, show_edges=None, cp={-6, -3, 160},
     if climits is None:
         cmin = 1.01 * np.min(signal)
         cmax = 1.01 * np.max(signal)
-        climits = {cmin, cmax}
+        climits = [cmin, cmax]
 
     # pygtgraph window initialization in 2D and 3D
     global window_list