+++ HISTORY.rst
@@ -4,13 +4,37 @@
 History
 =======
 
-0.0.1 (2014-10-06)
+0.2.1 (2015-10-14)
 ------------------
 
-Toolbox template release.
+Fix bug on pip installation.
+
+Update full documentation.
+
+
+0.2.0 (2015-10-05)
+------------------
+
+Adding functionalities to match the content of the Matlab GSP Box.
+
+First release of the PyGSP.
+
+
+0.1.0 (2015-06-02)
+------------------
+
+Main features of the box are present most of the graphs and filters can be used.
+
+The utils and operators modules also have most of their features implemented.
 
 
 0.0.2 (2015-04-19)
 ------------------
 
 Beginning of user tests.
+
+
+0.0.1 (2014-10-06)
+------------------
+
+Toolbox template release.